Fix: move BottomPlayerButton to domain module
GesturePreferences.kt (domain/player/service) referenced BottomPlayerButton
which still lived in the app module - domain modules can't depend on app
modules in this project's Gradle structure, causing a compile failure.

Moved BottomPlayerButton into domain/player/model/PlayerEnums.kt alongside
the other moved enums, and updated the 4 app-module call sites to import
it from its new location.
