+++ app/src/main/java/eu/kanade/tachiyomi/ui/player/PlayerViewModel.kt
@@ -15,6 +15,7 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
 import animiru.domain.player.interactor.TrackSelect
 import animiru.domain.player.model.ArtType
+import animiru.domain.player.model.BottomPlayerButton
 import animiru.domain.player.model.CustomKeyCodes
 import animiru.domain.player.model.PlayerOrientation
 import animiru.domain.player.model.SetAsArt
