+++ domain/src/main/java/animiru/domain/player/model/PlayerEnums.kt
@@ -245,3 +245,20 @@ enum class SubtitlesBorderStyle(
         }
     }
 }
+
+// AM (BOTTOM_PLAYER_BUTTONS) -->
+/**
+ * Buttons that can be shown in the customizable row along the bottom of the player.
+ * Order in [entries] is only the declaration order; the user's chosen order and
+ * selection are stored separately via [animiru.domain.player.service.GesturePreferences.bottomPlayerButtons].
+ */
+enum class BottomPlayerButton(val stringRes: StringResource) {
+    Lock(AYMR.strings.pref_bottom_button_lock),
+    Rotation(AYMR.strings.pref_bottom_button_rotation),
+    Speed(AYMR.strings.pref_bottom_button_speed),
+    Chapters(AYMR.strings.pref_bottom_button_chapters),
+    PictureInPicture(AYMR.strings.pref_bottom_button_pip),
+    AspectRatio(AYMR.strings.pref_bottom_button_aspect_ratio),
+    Screenshot(AYMR.strings.pref_bottom_button_screenshot),
+}
+// <-- AM (BOTTOM_PLAYER_BUTTONS)
